feat(sprint-3/phase-1): calculate document inventory statistics
Add analysis text normalization, RU/EN classification, heading projections,
single-document statistics, and deterministic batch inventory orchestration.

Refs: CORPUS-003

+++ src/rag_based_on_obsidian/corpus/inventory_entities.py
@@ -23,7 +23,7 @@ class TokenCount:
     """Token count produced by one named tokenizer baseline."""
 
     tokenizer_name: str
-    count: int
+    count: float
 
 
 @dataclass(frozen=True)
